feat: only use a single tailwind config
