--- apps/resume/tailwind.config.ts
@@ -1,3 +1,0 @@
-import baseConfig from '../../tailwind.config'
-
-export default baseConfig;
